Revert sessionId UPSERT client; trim path-catalog shims.
Drop client sessionId POST, session module, and UPSERT commit markers while keeping name-key lower-retry UX. Migrate imports to path-catalog/, remove deprecated shims and dead exports, refresh docs.
