--- js/puzzle-export-sim/load-path-catalog.js
@@ -1,2 +1,0 @@
-/** @deprecated Import from `./path-catalog/load-path-catalog.js` or `./path-catalog/index.js`. */
-export * from "./path-catalog/load-path-catalog.js";
