--- js/puzzle-export-sim/path-signature.js
@@ -1,2 +1,0 @@
-/** @deprecated Import from `./path-catalog/path-signature.js` or `./path-catalog/index.js`. */
-export * from "./path-catalog/path-signature.js";
