--- js/puzzle-export-sim/path-variant-catalog.js
@@ -1,2 +1,0 @@
-/** @deprecated Import from `./path-catalog/path-variant-catalog.js` or `./path-catalog/index.js`. */
-export * from "./path-catalog/path-variant-catalog.js";
